requirements.txt
customtkinter==5.2.0
pymongo==4.5.0
pillow==10.0.0
pyyaml==6.0.1
python-dotenv==1.0.0
ultralytics==8.0.196
loguru==0.7.0
cryptography==41.0.3
psutil==5.9.5
requests==2.31.0
opencv-python==4.8.0.76
numpy==1.24.3
pyautogui==0.9.54 